feat: call map-key w/ parent-schema information
This commit passes the parent-schema (ie. the map) to get the transformer
function.

This is useful if your map-keys wants to introspect the schema to determine
its behavior. An example usecase would be automatically remove the most
prevelant namespace of a key, and adding a string underscore encoding
for other namespaces. Eg. you can use this to write a json encoder that
does something like the following:

```clj
{:person/name "Bob" :person/age 36 :pet/name "Edgar"}
;; =>
{"name" "Bob" "age" 36 "pet_name" "Edgar"}
```