feat: :map ::m/rename-keys transformer metadata
Introduces the ability for map transformers to provide (at compile time
or run-time) information about key-ranames they perform by including the
`::m/rename-keys` key-map in metadata attached to either a run-time
value (ie. the data returned by the map transformer) or on the returned
interceptor functions themselves.

This allows map transformers to not break subsequent calls to child
value transformers.

Initially I tried to solve this generically with a `:leave` interceptor
in the map transformer, but this broke down with nested maps since
`:after` is currently executed as a prewalk.

It may be useful to also change the behavior of `:leave` to behave more
like a postwalk, but that requires more thought and is certainly a much
larger PR.