BREAKING: don't coerce map-of keywords to strings
Removes the default (and somewhat surprising) behaviour of coercing
keywords to strings before passing them to their transformers.

This had the unintended side-effect of making intentional keyword-like
schemas in map-of schemas have to be aware of how to coerce themselves
from both keywords and strings.

This behavior is only breaking for people writing their own custom
transformers with `map-of` keys. You now no longer need to handle both
strings and keyword inputs.

For end users using `mt/json-transformers` or `mt/string-transformers`,
adds a default `map-of` transformer which keeps the behavior that was
previously hard coded.

Closes #131