fix: type changing transformers
This commit introduces code and testing for two fixes:

1) It removes the check on the type of the data before calling the
`this-transformer` for the `tuple`, `map-of`, `map` and
`sequence`-related transformers. This is necessary to allow for encoded
/ decoded data to be in shapes that differ in type from their root
schema. Eg. a map schema that is encoded as a tuple of values (common in
financial candle data).

2) Introduces the ability for a `-chain` to take a tuple with a
`predicate` and a `sub-chain`. This allows schemas to specify groups of
child-transformations that depend on a certain predicate. This fixes an
error where post `this-transformer` functions would change their type to
a shape that would cause invoking the children to crash.