Add instrumentation and clj-kondo support for ClojureScript
Implements function instrumentation for ClojureScript code, attempting
to match as closely as possible to the Clojure implementation.

Adds separate namespaces to implement instrumentation support in cljs
because macros are required to dynamically change vars in cljs
(symbol literals are required at runtime).

malli.dev.cljs/start! does not feature a watch on the function schemas
atom because the var names need to be known at compile time in order
for a macro to produce valid cljs code. In cljs, hot code reloading
is the norm so this isn'ta major deficiency.