Allow custom options at `mx/defn`, output custom type to clj-kondo config and allow namespaced symbols to the `=>` macro
I was trying to use custom types in `mx/defn` and was not able to, so
here I'm just passing `:malli/options` metadata which has the
custom options (with some `:registry`). Also I'm checking if there is
some `prepost` instead of outputting `nil`.

`m/=>` was not working to namespaced symbols (maybe this was
intentional?), so this PR allows it to pass it, altough it has to be
fully qualified.

When emitting the configuration to `clj-kondo` when I have some custom
type, it was returning `:any`, but now it's outputing a type which the
user defined in the type properties at the `:clj-kondo/type` key.

Let me know if it makes sense o/