Add a more useful default error reporter for cljs instrumentation
Fipp, the pretty printer, is not designed with the browser console in
mind and the output when developing in cljs is not usable.
(Every print call invokes console.log which will result in a new output line per invocation,
instead of all on one line like in a terminal).
This commit allows navigating directly to the callsite where the error occurred
 via a JavaScript stacktrace.