Refactor cljs fn instrumentation to occur at runtime
Based on feedback from Thomas Heller this PR moves function instrumentation to
occur in the JS runtime instead of via macros which emit code to replace the JS functions.
The only part that happens at compile time is schema collection.
This results in a much closer match to the Clojure implementation.
One downside of this approach is that instrumentation will not work under
advanced compilation, but since its intended use is during development this tradeoff makes sense.