fix: explain :path of ref schemas
The path from a [:ref ::kw] or a [:ref #'Var] to the referred schema
is [0 0] due to the intermediate -pointer. See additions to
get-in-test. Thus the -explainer of a -ref-schema needs to add [0 0]
to the path, instead of the previous [0]. This mirrors the [0 0]
already present in the -walk of -ref-schema.

fixes #1106